requirements.txt
streamlit>=1.31.1,<2.0.0

langchain-core>=1.2.13,<2.0.0
langchain-community>=0.4.1,<1.0.0
langchain-openai==1.1.10
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1

faiss-cpu>=1.8.0,<2.0.0
sentence-transformers>=2.7.0,<3.0.0

python-dotenv>=1.0.1,<2.0.0
pypdf>=4.0.0,<5.0.0
langgraph>=0.2.0
google-search-results>=2.4.2
